修复 药品出入库“金额”输入小数报错问题

+++ src/main/java/me/zbl/app/domain/DrugOutFormDO.java
@@ -36,7 +36,7 @@ public class DrugOutFormDO {
   //  经办人
   private String manager;
   //  总金额
-  private int ammount;
+  private float ammount;
 
   public String getDrugId() {
     return drugId;
@@ -62,11 +62,11 @@ public class DrugOutFormDO {
     this.manager = manager;
   }
 
-  public int getAmmount() {
+  public float getAmmount() {
     return ammount;
   }
 
-  public void setAmmount(int ammount) {
+  public void setAmmount(float ammount) {
     this.ammount = ammount;
   }
